Re-initialize bneck variable for each beta. Tidy up code
Re-initializing the pdf for the bottleneck variable for each choice of
beta helps in revealing when the algorithm is stuck in a local
minimum.

Removed several superfluous functions; now we only iterate over the
code once to build a joint pdf, and we derive all other
distributions (conditional and marginal) from there.