Decouple EmpiricalBottleneck instantiation and BA algo run; update README
Also, some minor documentation fixes (giving more emphasis to 'X' and
'Y' rather than 'past' and 'future').

Now the BA algorithm is not automatically run as soon as an
EmpiricalBottleneck object is instantiated. This allows a user to
separate logically the part of the code where the analysis is
defined (the data is stored in the EB object together with the algo
settings) from that where the AB algorithm is actually run, which can
be computationally expensive. This is a crude version of the principle
adopted by e.g. scikit-learn, where one defines a classifier by
instantiating an object containing training data and settings, and
then runs the actual fit by calling its fit() method.