Lots of bugfixes, GIB/DIB seems fine now.
In decreasing order of importance:

- make sure unused bottleneck dimensions are dropped as soon as
possible. This is necessary to avoid various headaches with infinities
and NaNs due to stuff like dividing by zero when computing p(y|m), or
when computing the KL divergence.
- update initialization rules to make them more similar to the default
ones described in Strouse and Schwab 2016, when these are applicable.
- fix a bug with how monotonicity was enforced in the DIB plane, and
update the logic surrounding that bit to allow the user to choose what
type of monotonicity-ization they want (if any).
- add option passthrough to enable control of convergence criterion.
- remove nonzero absolute tolerance from convergence criterion.